Sửa lỗi Kết nối không hợp lý trên DB

+++ backend/bookCar/src/main/java/com/freelance/bookCar/models/product/ticket/Tourism.java
@@ -14,14 +14,10 @@ import lombok.NoArgsConstructor;
 @Entity
 public class Tourism {
     @Id
-    private Long id;
-
+    private Integer id;
     private String name;
     private String location;
     private String description;
     private double rating;
-
-
-
     // Getters and Setters
 }
